ApprovalRow - move renderRow/Status/Buttons logic from certification dashboard
and change isApproved to use .repository.pulp_labels.pipeline
